@@ -13,15 +13,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.0",
-    "@angular/common": "^7.1.0",
-    "@angular/compiler": "^7.1.0",
-    "@angular/core": "^7.1.0",
-    "@angular/forms": "^7.1.0",
-    "@angular/http": "^7.1.0",
-    "@angular/platform-browser": "^7.1.0",
-    "@angular/platform-browser-dynamic": "^7.1.0",
-    "@angular/router": "^7.1.0",
+    "@angular/animations": "^8.2.1",
+    "@angular/cdk": "^8.1.2",
+    "@angular/common": "^8.2.1",
+    "@angular/compiler": "^8.2.1",
+    "@angular/core": "^8.2.1",
+    "@angular/forms": "^8.2.1",
+    "@angular/platform-browser": "^8.2.1",
+    "@angular/platform-browser-dynamic": "^8.2.1",
+    "@angular/router": "^8.2.1",
     "@types/cors": "^2.8.4",
     "@types/express": "^4.16.0",
     "@types/hashids": "^1.0.30",
@@ -38,24 +38,24 @@
     "mongoose": "^5.3.1",
     "path": "^0.12.7",
     "primeflex": "^1.0.0-rc.1",
-    "primeicons": "^1.0.0-beta.10",
-    "primeng": "^6.1.3",
+    "primeicons": "^2.0.0",
+    "primeng": "^8.0.2",
     "rxjs": "^6.4.0",
     "tmp": "^0.1.0",
     "util.promisify": "^1.0.0",
-    "zone.js": "^0.8.19"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^7.1.0",
-    "@angular/compiler-cli": "^7.1.0",
-    "@angular/language-service": "^7.1.0",
+    "@angular/cli": "^8.2.1",
+    "@angular/compiler-cli": "^8.2.1",
+    "@angular/language-service": "^8.2.1",
     "@compodoc/compodoc": "^1.1.5",
     "@types/bluebird-global": "^3.5.11",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~10.0.0",
     "bluebird": "^3.5.4",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
@@ -66,7 +66,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~3.1.6",
-    "@angular-devkit/build-angular": "~0.6.0"
+    "typescript": "~3.5.3",
+    "@angular-devkit/build-angular": "~0.802.1"
   }
 }
